Move Header and Footer to separate Layout component
